leveraging on #106 and using Karabina's cyclotomic squaring for two series of 0 (#75) with a batch decompression, we can improve BLS12-381 final exponentiation.
```
benchmark                           old ns/op     new ns/op     delta
BenchmarkFinalExponentiation-32     424414        395896        -6.72%
```